misc: lis3dh_mot: Use REL events to report the acceleration
An ABS event is only sent if it differs from the previous one.
If the device is stationary we might omit several events,
especially if the sensitivity is low. This can cause issues if
userspace expects to receive a certain number of events for a
given time window. For example the Android framework polls the
accelerometer to determine if the device is stationary, but
given the use of ABS events, it often fails to collect enough
data. Switch to REL events to fix this.

Change-Id: Ia930dafa4cd7212a74f43506a56b1590395061b1